code/frontend/requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.2
frozendict==2.4.0
html5lib==1.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
lxml==5.1.0
MarkupSafe==2.1.5
multitasking==0.0.11
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.26.4
pandas==2.2.0
peewee==3.17.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
soupsieve==2.5
textblob==0.17.1
tqdm==4.66.1
tzdata==2023.4
urllib3==2.2.0
webencodings==0.5.1
werkzeug==3.0.1
yfinance==0.2.36
zipp==3.17.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.8
